requirements.txt
fastapi
uvicorn
strawberry-graphql
httpie